Access slice with given name instead of arbitrary property name
Before it was a magic property that depended on configuration in root reducer.
Now there is an implicit expectation that the root reducer place each slice
under it's name.

The `slice` property of each slice should be exported as part of the standard
structure.

Note that `slice?` will change to `name` in an upcoming release of rsk, but we
can keep it exported as slice to limit impact