Add a shuffle algorithm
Broke recipe slice into two seperate slices, one for ui and one for recipe
management. We will problably need a parent for poth of these to combine them
without cyclic deps, but this works for now.

The shuffle should move cards around based on a normal distribution generated
based on a hash of the recipe. On average, a card should move about one place
from it's natural order.